+++ Runtime/ISaveGameManager.cs
@@ -2,6 +2,7 @@
 // See LICENSE in the project root for license information.
 
 using System;
+using UniSharper.Data.SaveGame.Providers;
 
 namespace UniSharper.Data.SaveGame
 {
@@ -31,8 +32,8 @@ namespace UniSharper.Data.SaveGame
         /// Initializes save game manager.
         /// </summary>
         /// <param name="storePath">The store path where data to save.</param>
-        /// <param name="cryptoProvider">The crypto provider.</param>
-        /// <param name="compressionProvider">The compression provider.</param>
+        /// <param name="cryptoProvider">The crypto provider. Sets <c>null</c> to use <see cref="AesCryptoProvider"/>. </param>
+        /// <param name="compressionProvider">The compression provider. Sets <c>null</c> to use <see cref="DeflateCompressionProvider"/>. </param>
         void Initialize(string storePath = null, ICryptoProvider cryptoProvider = null, ICompressionProvider compressionProvider = null);
 
         /// <summary>